+++ src/core/CardViewer/CVGridCardObject.gd
@@ -14,7 +14,7 @@ func _ready() -> void:
 
 func setup(card) -> Card:
 	if typeof(card) == TYPE_STRING:
-		display_card = cfc.instance_card(card)		
+		display_card = cfc.instance_card(card)
 	else:
 		display_card = card
 		display_card.position = Vector2(0,0)